+++ tests/test_testing.py
@@ -2,6 +2,8 @@ import os
 import os.path
 import tempfile
 import gzip
+import threading
+import time
 
 import pytest
 
@@ -54,6 +56,62 @@ def test_non_text_file_can_be_stored(samba_mock: SMBConnectionMock):
         )
 
 
+def test_async_retrieval(samba_mock: SMBConnectionMock):
+    connection = pyndows.connect(
+        "TestComputer", "127.0.0.1", 80, "TestDomain", "TestUser", "TestPassword"
+    )
+    with tempfile.TemporaryDirectory() as temp_dir:
+        with gzip.open(os.path.join(temp_dir, "local_file"), mode="w") as distant_file:
+            distant_file.write(b"Test Content Move")
+
+        def add_with_delay(delay: int):
+            time.sleep(delay)
+            pyndows.move(
+                connection,
+                "TestShare",
+                "TestFilePath",
+                os.path.join(temp_dir, "local_file"),
+            )
+
+        threading.Thread(target=add_with_delay, args=(2,)).start()
+
+        assert (
+            gzip.decompress(
+                samba_mock.stored_files.try_get(
+                    ("TestShare", "TestFilePath"), timeout=3
+                )
+            )
+            == b"Test Content Move"
+        )
+
+
+def test_async_retrieval_timeout(samba_mock: SMBConnectionMock):
+    connection = pyndows.connect(
+        "TestComputer", "127.0.0.1", 80, "TestDomain", "TestUser", "TestPassword"
+    )
+    with tempfile.TemporaryDirectory() as temp_dir:
+        with gzip.open(os.path.join(temp_dir, "local_file"), mode="w") as distant_file:
+            distant_file.write(b"Test Content Move")
+
+        def add_with_delay(delay: int):
+            time.sleep(delay)
+            pyndows.move(
+                connection,
+                "TestShare",
+                "TestFilePath",
+                os.path.join(temp_dir, "local_file"),
+            )
+
+        threading.Thread(target=add_with_delay, args=(2,)).start()
+
+        with pytest.raises(TimeoutError) as exception_info:
+            samba_mock.stored_files.try_get(("TestShare", "TestFilePath"))
+        assert (
+            str(exception_info.value)
+            == "('TestShare', 'TestFilePath') could not be found within 1 seconds."
+        )
+
+
 def test_file_retrieval_using_path(samba_mock: SMBConnectionMock):
     connection = pyndows.connect(
         "TestComputer", "127.0.0.1", 80, "TestDomain", "TestUser", "TestPassword"
@@ -73,3 +131,73 @@ def test_file_retrieval_using_path(samba_mock: SMBConnectionMock):
         )
         with gzip.open(os.path.join(temp_dir, "local_file_retrieved")) as local_file:
             assert local_file.read() == b"Test Content"
+
+
+def test_file_retrieval_using_str_content(samba_mock: SMBConnectionMock):
+    connection = pyndows.connect(
+        "TestComputer", "127.0.0.1", 80, "TestDomain", "TestUser", "TestPassword"
+    )
+    with tempfile.TemporaryDirectory() as temp_dir:
+        samba_mock.files_to_retrieve[("TestShare", "TestFilePath")] = "data"
+
+        pyndows.get(
+            connection,
+            "TestShare",
+            "TestFilePath",
+            os.path.join(temp_dir, "local_file_retrieved"),
+        )
+        with open(os.path.join(temp_dir, "local_file_retrieved"), "rt") as local_file:
+            assert local_file.read() == "data"
+
+
+def test_file_retrieval_using_bytes_content(samba_mock: SMBConnectionMock):
+    connection = pyndows.connect(
+        "TestComputer", "127.0.0.1", 80, "TestDomain", "TestUser", "TestPassword"
+    )
+    with tempfile.TemporaryDirectory() as temp_dir:
+        bytes_content_file_path = os.path.join(temp_dir, "local_file")
+        with gzip.open(bytes_content_file_path, mode="w") as distant_file:
+            distant_file.write(b"Test Content")
+        samba_mock.files_to_retrieve[("TestShare", "TestFilePath")] = open(
+            bytes_content_file_path, "rb"
+        ).read()
+
+        pyndows.get(
+            connection,
+            "TestShare",
+            "TestFilePath",
+            os.path.join(temp_dir, "local_file_retrieved"),
+        )
+        with gzip.open(os.path.join(temp_dir, "local_file_retrieved")) as local_file:
+            assert local_file.read() == b"Test Content"
+
+
+def test_retrieval_of_stored_non_text_file(samba_mock: SMBConnectionMock):
+    connection = pyndows.connect(
+        "TestComputer", "127.0.0.1", 80, "TestDomain", "TestUser", "TestPassword"
+    )
+    with tempfile.TemporaryDirectory() as temp_dir:
+        with gzip.open(os.path.join(temp_dir, "local_file"), mode="w") as distant_file:
+            distant_file.write(b"Test Content Move")
+
+        pyndows.move(
+            connection,
+            "TestShare",
+            "TestFilePath",
+            os.path.join(temp_dir, "local_file"),
+        )
+
+        pyndows.get(
+            connection,
+            "TestShare",
+            "TestFilePath",
+            os.path.join(temp_dir, "local_file_retrieved"),
+        )
+
+        with gzip.open(os.path.join(temp_dir, "local_file_retrieved")) as local_file:
+            assert local_file.read() == b"Test Content Move"
+
+        assert (
+            gzip.decompress(samba_mock.stored_files[("TestShare", "TestFilePath")])
+            == b"Test Content Move"
+        )