+++ CHANGELOG.md
@@ -6,9 +6,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [3.2.0] - 2020-01-22
+### Fixed
+- Allow to store bytes content using Mock.
+- Mock is now able to retrieve a file that was previously stored.
+
+### Added
+- Mock stored files can now be retrieved with the help of a timeout via `try_get` method.
+
 ## [3.1.0] - 2019-12-03
 ### Added
 - Initial release.
 
-[Unreleased]: https://github.com/Colin-b/pyndows/compare/v3.1.0...HEAD
+[Unreleased]: https://github.com/Colin-b/pyndows/compare/v3.2.0...HEAD
+[3.2.0]: https://github.com/Colin-b/pyndows/compare/v3.1.0...v3.2.0
 [3.1.0]: https://github.com/Colin-b/pyndows/releases/tag/v3.1.0